+++ src/main/java/com/github/jlangch/venice/impl/docgen/DocGenerator.java
@@ -1038,6 +1038,7 @@ public class DocGenerator {
 		generic.addItem(getDocItem("fn"));
 		generic.addItem(getDocItem("loop"));
 		generic.addItem(getDocItem("recur"));
+		generic.addItem(getDocItem("set!"));
 		generic.addItem(getDocItem("try", true, true));   // FIXME: try
 		generic.addItem(getDocItem("try-with", true, true));